+++ examples/exmp054_casscf_fcidump/job.py
@@ -5,12 +5,12 @@ import sys
 from pathlib import Path
 
 from opi.core import Calculator
-from opi.input.simple_keywords import AuxBasisSet, BasisSet, Scf, Wft
+from opi.input.blocks import BlockCasscf, BlockOutput
 from opi.input.structures import Structure
 from opi.output.core import Output
 
 
-def run_exmp029(
+def run_exmp054(
     structure: Structure | None = None, working_dir: Path | None = Path("RUN")
 ) -> Output:
     # > recreate the working dir
@@ -23,9 +23,8 @@ def run_exmp029(
 
     calc = Calculator(basename="job", working_dir=working_dir)
     calc.structure = structure
-    calc.input.add_simple_keywords(
-        Scf.NOAUTOSTART, Wft.OO_RI_MP2, BasisSet.DEF2_QZVPP, AuxBasisSet.DEF2_QZVPP_C
-    )
+
+    calc.input.add_blocks(BlockCasscf(nel=2, norb=2), BlockOutput(dumpactints=True))
     calc.input.ncores = 4
 
     calc.write_input()
@@ -36,19 +35,29 @@ def run_exmp029(
         print(f"ORCA calculation failed, see output file: {output.get_outfile()}")
         print(output.error_message())
         sys.exit(1)
-    # << END OF IF
+    if not output.casscf_converged():
+        print(f"ORCA calculation failed, see output file: {output.get_outfile()}")
+        print(output.error_message())
+        sys.exit(1)
 
     # > Parse JSON files
     output.parse()
 
-    ngeoms = len(output.results_properties.geometries)
-    print("N GEOMETRIES")
-    print(ngeoms)
-    print("HF ENERGY")
-    print(output.results_properties.geometries[0].energy[1].refenergy[0][0])
-    print("CORRELATION ENERGY")
-    print(output.results_properties.geometries[0].energy[1].correnergy[0][0])
-    print("TOTAL ENERGY")
-    print(output.results_properties.geometries[0].energy[1].totalenergy[0][0])
+    print("FINAL SINGLE POINT ENERGY")
+    print(output.results_properties.geometries[0].single_point_data.finalenergy)
+    print("CASSCF energy")
+    print(output.results_properties.geometries[0].energy[0].totalenergy[0][0])
+
+    # > retrieve the path to the fcidump file
+    fcidump_file = output.get_outfile().with_suffix(".fcidump")
+
+    # > Read and print the file
+    print("FCIDUMP file:")
+    with open(fcidump_file, "r") as f:
+        print(f.read())
 
     return output
+
+
+if __name__ == "__main__":
+    run_exmp054()